# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "description": "API for E-Commerce Application",
    "title": "E-Commerce API "
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT": {
        "in": "header",
        "name": "Authorization",
        "type": "apiKey"
      }
    }
  },
  "paths": {
    "/api/users/signup": {
      "post": {
        "tags": ["Users"],
        "summary": "SignUp",
        "description": "User SignUp to Create Account",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Incorrect Credentials"
          }
        }
      }
    },
    "/api/users/signin": {
      "post": {
        "tags": ["Users"],
        "summary": "Login",
        "description": "User Login to get token",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Incorrect Credentials"
          }
        }
      }
    },
    "/api/users/resetPassword": {
      "post": {
        "tags": ["Users"],
        "summary": "Reset Password",
        "description": "Reset User Password using Email and Old Password",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "newPassword": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Incorrect Credentials"
          }
        }
      }
    },

    "/api/products": {
      "get": {
        "tags": ["Products"],
        "summary": "Get All Products",
        "description": "User will get all products",
        "security": [
          {
            "JWT": {}
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": ["Products"],
        "summary": "Add a Product",
        "description": "User can add a product",
        "security": [
          {
            "JWT": {}
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "desc": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number"
                  },
                  "categories": {
                    "type": "string"
                  },
                  "inStock": {
                    "type": "number"
                  },
                  "sizes": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/products/{id}": {
      "get": {
        "tags": ["Products"],
        "summary": "Get a Product by id",
        "description": "Retrieve a specific product by its unique ID.",
        "security": [
          {
            "JWT": {}
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the product to retrieve",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/products/rateproduct": {
      "post": {
        "tags": ["Products"],
        "summary": "Rate a Product between 1 to 5",
        "description": "User can rate a product between 1 to 5",
        "security": [
          {
            "JWT": {}
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ratings": {
                    "type": "number",
                    "minimum": 1,
                    "maximum": 5
                  },
                  "productId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/products/filter":{
      "get": {
        "tags": ["Products"],
        "summary": "Filter Producta",
        "description": "User can filter product using minimum Price and Categories",
        "security": [
          {
            "JWT": {}
          }
        ],
        "parameters": [
          {
            "name": "minPrice",
            "in": "query",
            "required": true,
            "description": "Minimum Price of the product",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "categories",
            "in": "query",
            "required": true,
            "description": "Id of Categories of the product seperated by comma(,) in string format",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },

    "/api/cartItems/":{
      "get": {
        "tags": ["Cart"],
        "summary": "Get CartItems of User",
        "description": "User will get all cart items",
        "security": [
          {
            "JWT": {}
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": ["Cart"],
        "summary": "Add Product to Cart with quantity",
        "description": "User can add product to cart with quantity",
        "security": [
          {
            "JWT": {}
          }
        ],
        "requestBody":{
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "productId": {
                    "type": "string"
                  },
                  "quantity": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/cartItems/{id}":{
      "delete": {
        "tags": ["Cart"],
        "summary": "Delete CartItem using CartItem's Id",
        "description": "User can delete cart item using cart item's id",
        "security": [
          {
            "JWT": {}
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the cart item to delete",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },

    "/api/likes/":{
      "get": {
        "tags": ["Like"],
        "summary": "Get All Likes",
        "description": "User will get all likes",
        "security": [
          {
            "JWT": {}
          }
        ],
        "parameters": [
          {
            "name": "typeId",
            "in": "query",
            "required": true,
            "description": "ID of the product/category to like",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": true,
            "description": "Type - (Product/Category)",
            "schema": {
              "type": "string",
              "enum": ["Product", "Category"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": ["Like"],
        "summary": "Like to Product",
        "description": "User can like a product",
        "security": [
          {
            "JWT": {}
          }
        ],
        "requestBody":{
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "typeId": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string",
                    "enum": ["Product", "Category"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },

    "/api/categories/":{
      "get": {
        "tags": ["Categories"],
        "summary": "Get All Categories",
        "description": "User will get all categories",
        "security": [
          {
            "JWT": {}
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    }

  }
}